git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@103068 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 3fa1b57a8f31572b8bce2c9349217474d0cd05f4

+++ Online/Online/ROMon/src/ROMonBridge.cpp
@@ -65,7 +65,7 @@ namespace ROMon {
     /// Prefix for resulting service names
     std::string     m_prefix;
     /// Printout level
-    int             m_print;
+    long            m_print;
 
     struct ClusterAdder {
       ROMonBridge* bridge;
